+++ packages/story-editor/src/app/story/useStoryReducer/reducers/deleteElements.js
@@ -17,7 +17,7 @@
 /**
  * Internal dependencies
  */
-import { getAllProducts, intersect } from './utils';
+import { intersect } from './utils';
 
 /**
  * Delete elements by the given list of ids.
@@ -106,13 +106,10 @@ function deleteElements(state, { elementIds }) {
     ? state.selection.filter((id) => !validDeletionIds.includes(id))
     : state.selection;
 
-  const products = getAllProducts(newPages);
-
   return {
     ...state,
     pages: newPages,
     selection: newSelection,
-    products,
   };
 }
