+++ packages/story-editor/src/app/story/useStoryReducer/reducers/deletePage.js
@@ -15,11 +15,6 @@
  */
 
 /**
- * Internal dependencies
- */
-import { getAllProducts } from './utils';
-
-/**
  * Delete page by id or delete current page if no id given.
  *
  * If another page than current page is deleted, it will remain current page.
@@ -62,14 +57,11 @@ function deletePage(state, { pageId }) {
     newCurrent = newPages[newCurrentIndex].id;
   }
 
-  const products = getAllProducts(newPages);
-
   return {
     ...state,
     pages: newPages,
     current: newCurrent,
     selection: [],
-    products,
   };
 }
