+++ packages/story-editor/src/app/story/useStoryReducer/reducers/restore.js
@@ -15,11 +15,6 @@
  */
 
 /**
- * Internal dependencies
- */
-import { getAllProducts } from './utils';
-
-/**
  * Restore internal state completely from given state.
  *
  * Some validation is performed:
@@ -36,13 +31,9 @@ import { getAllProducts } from './utils';
  * @param {Array} payload.selection Selection.
  * @param {Object} payload.story Story object.
  * @param {Object} payload.capabilities Capabilities object.
- * @param {Array} payload.products Array of products.
  * @return {Object} New state
  */
-function restore(
-  state,
-  { pages, current, selection, story, capabilities, products }
-) {
+function restore(state, { pages, current, selection, story, capabilities }) {
   if (!Array.isArray(pages) || pages.length === 0) {
     return state;
   }
@@ -54,11 +45,9 @@ function restore(
     ? oldCurrent
     : pages[0].id;
   const newSelection = Array.isArray(selection) ? selection : [];
-  const newProducts = products ? products : getAllProducts(pages);
 
   return {
     pages,
-    products: newProducts,
     current: newCurrent,
     selection: newSelection,
     story: newStory,
