+++ packages/story-editor/src/app/story/useStoryReducer/useStoryReducer.js
@@ -30,7 +30,6 @@ const INITIAL_STATE = {
   pages: [],
   capabilities: {},
   current: null,
-  products: [],
   selection: [],
   story: {},
   animationState: STORY_ANIMATION_STATE.RESET,
